+++ docs/changelog.rst
@@ -6,6 +6,8 @@ Changelog
 `Unreleased`_
 -------------
 
+- Allow ``block_network.allowed_hosts`` configuration via ``vcr_config`` fixture. `#82`_
+
 `0.12.0`_ - 2021-07-08
 ----------------------
 
@@ -194,6 +196,7 @@ Added
 .. _0.3.0: https://github.com/kiwicom/pytest-recording/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/pytest-recording/compare/v0.1.0...v0.2.0
 
+.. _#82: https://github.com/kiwicom/pytest-recording/pull/82
 .. _#69: https://github.com/kiwicom/pytest-recording/issues/69
 .. _#68: https://github.com/kiwicom/pytest-recording/issues/68
 .. _#64: https://github.com/kiwicom/pytest-recording/issues/64